# Try: Rewrite Block Rename UI with hooks