--- a/packages/block-editor/src/style.scss
+++ b/packages/block-editor/src/style.scss
@@ -20,6 +20,7 @@
 @import "./components/block-switcher/style.scss";
 @import "./components/block-types-list/style.scss";
 @import "./components/block-pattern-setup/style.scss";
+@import "./components/block-rename-modal/style.scss";
 @import "./components/block-variation-transforms/style.scss";
 @import "./components/border-radius-control/style.scss";
 @import "./components/colors-gradients/style.scss";
@@ -56,7 +57,6 @@
 @import "./hooks/padding.scss";
 @import "./hooks/position.scss";
 @import "./hooks/typography.scss";
-@import "./hooks/block-rename-ui.scss";
 
 @import "./components/block-toolbar/style.scss";
 @import "./components/inserter/style.scss";