--- a/packages/block-editor/src/style.scss
+++ b/packages/block-editor/src/style.scss
@@ -20,7 +20,7 @@
 @import "./components/block-switcher/style.scss";
 @import "./components/block-types-list/style.scss";
 @import "./components/block-pattern-setup/style.scss";
-@import "./components/block-rename-modal/style.scss";
+@import "./components/block-rename/style.scss";
 @import "./components/block-variation-transforms/style.scss";
 @import "./components/border-radius-control/style.scss";
 @import "./components/colors-gradients/style.scss";