Closes ##55612

## What?

This PR uses a hook to rewrite a series of processes to solve the problem that the Rename menu is not displayed in the popover menu of blocks that are not selected in the List View.

## Why?

Thanks to #54426, most blocks now support renaming. However, this resulted in a performance drop, so #55250 changed the component to only render when the block is selected.

This is correct behavior in the Advanced panel of the block sidebar, but in the List View I think the Rename menu should be visible like any other item, even if no block is selected. 

## How?

In trunk, we render a text field in the Advanced panel via the `editor.BlockEdit` hook and inject the Rename modal into the `BlockSettingsMenuControls` slot.

In this PR, the `editor.BlockEdit` hook only renders the text field in the Advanced panel.

Instead, the menu item to display the rename modal UI is added to the `BlockSettingsMenuControls` slot itself. This implementation and file structure is very similar to [the Block Locking implementation](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-editor/src/components/block-lock).

This allows renaming even if the block is not selected in the list view, and there should be no performance degradation.

## Testing Instructions

- You should be able to rename blocks from the Advanced panel.
- You should be able to rename blocks from the List View even if no blocks are selected.

### Perf test

To see locally how it affects performance, run the following commands on both the trunk branch and this PR:

```
npm run test:performance -- -g "Selecting blocks"
```

Below are the average values of three tests run on both branches. This value may depend on your local environment, but all metrics should be approximately the same.

#### This PR

```
┌──────────┬────────────┐
│ (index)  │   value    │
├──────────┼────────────┤
│  focus   │ '80.06 ms' │
│ minFocus │ '69.7 ms'  │
│ maxFocus │ '98.35 ms' │
└──────────┴────────────┘
```

#### trunk

```
┌──────────┬────────────┐
│ (index)  │   value    │
├──────────┼────────────┤
│  focus   │ '85.46 ms' │
│ minFocus │ '86.9 ms'  │
│ maxFocus │ '93.97 ms' │
└──────────┴────────────┘
```