@t-hamano 2023-10-26
I found PR #55604 for the `wp/6.4` branch and learned that it has almost the same approach. Therefore, I would like to close this PR.